+++ eng/common/scripts/Helpers/Metadata-Helpers.ps1
@@ -101,7 +101,7 @@ function compare-and-merge-metadata ($original, $updated) {
   }
   $originalTable = ConvertFrom-StringData -StringData $original -Delimiter ":"
   foreach ($key in $originalTable.Keys) {
-    if (!($updated.ContainsKey($key))) {
+    if (!($updated.Contains($key))) {
       Write-Warning "New metadata missed the entry: $key. Adding back."
       $updateMetdata += "$key`: $($originalTable[$key])`r`n"
     }
@@ -109,28 +109,13 @@ function compare-and-merge-metadata ($original, $updated) {
   return $updateMetdata
 }
 
-function GenerateDocsMsMetadata($originalMetadata, $language, $languageDisplayName, $serviceName, $tenantId, $clientId, $clientSecret, $msService) 
+function GenerateDocsMsMetadata($originalMetadata, $language, $languageDisplayName, $serviceName, $author, $msAuthor, $msService) 
 {
   $langTitle = "Azure $serviceName SDK for $languageDisplayName"
   $langDescription = "Reference for Azure $serviceName SDK for $languageDisplayName"
-  # Github url for source code: e.g. https://github.com/Azure/azure-sdk-for-js
-  $serviceBaseName = $serviceName.ToLower().Replace(' ', '').Replace('/', '-')
-  $author = GetPrimaryCodeOwner -TargetDirectory "/sdk/$serviceBaseName/"
-  $msauthor = ""
-  if (!$author) {
-    LogError "Cannot fetch the author from CODEOWNER file."
-  }
-  elseif ($TenantId -and $ClientId -and $ClientSecret) {
-    $msauthor = GetMsAliasFromGithub -TenantId $tenantId -ClientId $clientId -ClientSecret $clientSecret -GithubUser $author
-  }
-  # Default value
-  if (!$msauthor) {
-    LogError "No ms.author found for $author. "
-    $msauthor = $author
-  }
   $date = Get-Date -Format "MM/dd/yyyy"
 
-  $metadataTable = @{
+  $metadataTable = [ordered]@{
     "title"= $langTitle
     "description"= $langDescription
     "author"= $author