+++ eng/common/scripts/stress-testing/stress-test-deployment-lib.ps1
@@ -5,6 +5,19 @@ $FailedCommands = New-Object Collections.Generic.List[hashtable]
 
 . (Join-Path $PSScriptRoot "../Helpers" PSModule-Helpers.ps1)
 
+$limitRangeSpec = @"
+apiVersion: v1
+kind: LimitRange
+metadata:
+  name: default-resource-request
+spec:
+  limits:
+  - defaultRequest:
+      cpu: 100m
+      memory: 100Mi
+    type: Container
+"@
+
 # Powershell does not (at time of writing) treat exit codes from external binaries
 # as cause for stopping execution, so do this via a wrapper function.
 # See https://github.com/PowerShell/PowerShell-RFC/pull/277
@@ -40,8 +53,12 @@ function Login([string]$subscription, [string]$clusterGroup, [switch]$pushImages
     $cluster = RunOrExitOnFailure az aks list -g $clusterGroup --subscription $subscription -o json
     $clusterName = ($cluster | ConvertFrom-Json).name
 
-    $kubeContext = (RunOrExitOnFailure kubectl config view -o json) | ConvertFrom-Json
-    $defaultNamespace = $kubeContext.contexts.Where({ $_.name -eq $clusterName }).context.namespace
+    $kubeContext = (RunOrExitOnFailure kubectl config view -o json) | ConvertFrom-Json -AsHashtable
+    $defaultNamespace = $null
+    $targetContext = $kubeContext.contexts.Where({ $_.name -eq $clusterName }) | Select -First 1
+    if ($targetContext -ne $null -and $targetContext.PSObject.Properties.Name -match "namespace") {
+        $defaultNamespace = $targetContext.context.namespace
+    }
 
     RunOrExitOnFailure az aks get-credentials `
         -n "$clusterName" `
@@ -145,8 +162,10 @@ function DeployStressTests(
             -login:$login
     }
 
-    Write-Host "Releases deployed by $deployer"
-    Run helm list --all-namespaces -l deployId=$deployer
+    if ($FailedCommands.Count -lt $pkgs.Count) {
+        Write-Host "Releases deployed by $deployer"
+        Run helm list --all-namespaces -l deployId=$deployer
+    }
 
     if ($FailedCommands) {
         Write-Warning "The following commands failed:"
@@ -187,6 +206,9 @@ function DeployStressPackage(
     Write-Host "Creating namespace $($pkg.Namespace) if it does not exist..."
     kubectl create namespace $pkg.Namespace --dry-run=client -o yaml | kubectl apply -f -
     if ($LASTEXITCODE) {exit $LASTEXITCODE}
+    Write-Host "Adding default resource requests to namespace/$($pkg.Namespace)"
+    $limitRangeSpec | kubectl apply -n $pkg.Namespace -f -
+    if ($LASTEXITCODE) {exit $LASTEXITCODE}
 
     $dockerBuildConfigs = @()
     
@@ -201,6 +223,9 @@ function DeployStressPackage(
                 $dockerFilePath = "$($pkg.Directory)/Dockerfile"
             }
             $dockerFilePath = [System.IO.Path]::GetFullPath($dockerFilePath).Trim()
+            if (!(Test-Path $dockerFilePath)) {
+                continue
+            }
 
             if ("imageBuildDir" -in $scenario.keys) {
                 $dockerBuildDir = Join-Path $pkg.Directory $scenario.imageBuildDir
@@ -212,7 +237,7 @@ function DeployStressPackage(
         }
     }
     if ($pkg.Dockerfile -or $pkg.DockerBuildDir) {
-        throw "The chart.yaml docker config is depracated, please use the scenarios matrix instead."
+        throw "The chart.yaml docker config is deprecated, please use the scenarios matrix instead."
     }
     
 
@@ -249,9 +274,10 @@ function DeployStressPackage(
             }
         }
         $genVal.scenarios = @( foreach ($scenario in $genVal.scenarios) {
-            $dockerPath = Join-Path $pkg.Directory $scenario.image
-            if ("image" -notin $scenario) {
-                $dockerPath = $dockerFilePath
+            $dockerPath = if ("image" -notin $scenario) {
+                $dockerFilePath
+            } else {
+                Join-Path $pkg.Directory $scenario.image
             }
             if ([System.IO.Path]::GetFullPath($dockerPath) -eq $dockerFilePath) {
                 $scenario.imageTag = $imageTag
@@ -263,26 +289,36 @@ function DeployStressPackage(
     }
 
     Write-Host "Installing or upgrading stress test $($pkg.ReleaseName) from $($pkg.Directory)"
-    Run helm upgrade $pkg.ReleaseName $pkg.Directory `
-        -n $pkg.Namespace `
-        --install `
-        --set stress-test-addons.env=$environment `
-        --values generatedValues.yaml
+    $result = (Run helm upgrade $pkg.ReleaseName $pkg.Directory `
+                -n $pkg.Namespace `
+                --install `
+                --set stress-test-addons.env=$environment `
+                --values (Join-Path $pkg.Directory generatedValues.yaml)) 2>&1
+
     if ($LASTEXITCODE) {
-        # Issues like 'UPGRADE FAILED: another operation (install/upgrade/rollback) is in progress'
-        # can be the result of cancelled `upgrade` operations (e.g. ctrl-c).
-        # See https://github.com/helm/helm/issues/4558
-        Write-Warning "The issue may be fixable by first running 'helm rollback -n $($pkg.Namespace) $($pkg.ReleaseName)'"
-        return
+        # Error: UPGRADE FAILED: create: failed to create: Secret "sh.helm.release.v1.stress-test.v3" is invalid: data: Too long: must have at most 1048576 bytes
+        # Error: UPGRADE FAILED: create: failed to create: Request entity too large: limit is 3145728
+        if ($result -like "*Too long*" -or $result -like "*Too large*") {
+          $result
+          Write-Warning "*** Ensure any files or directories not part of the stress config are added to .helmignore ***"
+          Write-Warning "*** See https://helm.sh/docs/chart_template_guide/helm_ignore_file/ ***"
+          return
+        } else {
+          # Issues like 'UPGRADE FAILED: another operation (install/upgrade/rollback) is in progress'
+          # can be the result of cancelled `upgrade` operations (e.g. ctrl-c).
+          # See https://github.com/helm/helm/issues/4558
+          Write-Warning "The issue may be fixable by first running 'helm rollback -n $($pkg.Namespace) $($pkg.ReleaseName)'"
+          return
+        }
     }
 
     # Helm 3 stores release information in kubernetes secrets. The only way to add extra labels around
     # specific releases (thereby enabling filtering on `helm list`) is to label the underlying secret resources.
     # There is not currently support for setting these labels via the helm cli.
-    $helmReleaseConfig = kubectl get secrets `
-        -n $pkg.Namespace `
-        -l status=deployed,name=$($pkg.ReleaseName) `
-        -o jsonpath='{.items[0].metadata.name}'
+    $helmReleaseConfig = RunOrExitOnFailure kubectl get secrets `
+                                                -n $pkg.Namespace `
+                                                -l "status=deployed,name=$($pkg.ReleaseName)" `
+                                                -o jsonpath='{.items[0].metadata.name}'
 
     Run kubectl label secret -n $pkg.Namespace --overwrite $helmReleaseConfig deployId=$deployId
 }