+++ sdk/cognitiveservices/Azure.ResourceManager.CognitiveServices/CHANGELOG.md
@@ -1,10 +1,20 @@
 # Release History
 
-## 1.1.0 (Unreleased)
+## 1.2.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
 
 ### Other Changes
 
-- Switch to API version `2022-10-01`
+## 1.1.0 (2022-11-02)
+
+### Other Changes
+
+- Upgraded API version to 2022-10-01
 
 ## 1.0.0 (2022-08-29)