+++ sdk/textanalytics/Azure.AI.TextAnalytics/CHANGELOG.md
@@ -9,6 +9,13 @@
 - Added `WellKnownFhirVersion` and `HealthcareDocumentType` enums.
 - Added `TextAnalyticsActions.ExtractSummaryActions` to perform extractive summarization in a batch of actions.
 - Added `TextAnalyticsClient.StartExtractSummary` and `StartExtractSummaryAsync` to perform extractive summarization on a collection of documents.
+- Added `TextAnalyticsActions.AbstractSummaryActions` to perform abstractive summarization in a batch of actions.
+- Added `TextAnalyticsClient.StartAbstractSummary` and `StartAbstractSummaryAsync` to perform abstractive summarization on a collection of documents.
+- Added `Script` property to `DetectedLanguage`.
+- Added `ScriptKind` enum.
+- Added the `CategorizedEntity.Resolutions` property.
+- Added the following classes: `AgeResolution`, `AreaResolution`, `BaseResolution`, `BooleanResolution`, `CurrencyResolution`, `DateTimeResolution`, `InformationResolution`, `LengthResolution`, `NumberResolution`, `NumericRangeResolution`, `OrdinalResolution`, `SpeedResolution`, `TemperatureResolution`, `TemporalSpanResolution` `VolumeResolution`, and `WeightResolution`.
+- Added the following enums: `AgeUnit`, `AreaUnit`, `DateTimeSubKind`, `InformationUnit`,`LengthUnit`, `NumberKind`, `RangeKind`, `RelativeTo`, `SpeedUnit`, `TemperatureUnit`,`TemporalModifier`, `VolumeUnit`, and `WeightUnit`.
 
 ### Breaking Changes